requirements.txt
fastapi
uvicorn
pyodbc
pydantic
python-jose[cryptography]
passlib[bcrypt]
